Tests and refactoring

See merge request informatik/qe/swess23/group2/g2t4!32

+++ backend/src/main/java/at/qe/backend/services/UserxService.java
@@ -1,21 +1,23 @@
 package at.qe.backend.services;
 
 import at.qe.backend.configs.WebSecurityConfig;
+import at.qe.backend.exceptions.Userx.LastAdminException;
+import at.qe.backend.exceptions.Userx.UserAlreadyExistsException;
+import at.qe.backend.exceptions.Userx.UserDoesNotExistException;
 import at.qe.backend.models.UserRole;
 import at.qe.backend.models.Userx;
 import at.qe.backend.repositories.UserxRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
-import org.springframework.web.server.ResponseStatusException;
+import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.Date;
-import java.util.Set;
+import java.util.HashSet;
 
 /**
  * Service for accessing and manipulating user data.
@@ -23,12 +25,14 @@ import java.util.Set;
  * This class is part of the skeleton project provided for students of the
  * course "Software Architecture" offered by Innsbruck University.
  */
-@Component
+@Service
 @Scope("application")
 public class UserxService {
 
-    @Autowired
-    private UserxRepository userRepository;
+    private final UserxRepository userRepository;
+    public UserxService(UserxRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
     /**
      * Returns a collection of all users.
@@ -63,14 +67,12 @@ public class UserxService {
         if (user.isNew()) {
             user.setCreateDate(new Date());
             user.setCreateUserUsername(getAuthenticatedUsername());
-            if (user.getRoles().isEmpty()) {
-                user.setRoles(Set.of(UserRole.USER));
-            }
         } else {
             user.setUpdateDate(new Date());
             user.setUpdateUserUsername(getAuthenticatedUsername());
         }
-        return userRepository.save(user);
+        user = userRepository.save(user);
+        return user;
     }
 
     /**
@@ -79,13 +81,14 @@ public class UserxService {
      * @param user the user to delete
      */
     @PreAuthorize("hasAuthority('ADMIN')")
-    public boolean deleteUser(Userx user) {
+    public void deleteUser(Userx user) throws UserDoesNotExistException, LastAdminException {
+        if (!userRepository.existsByUsername(user.getUsername())) {
+            throw new UserDoesNotExistException();
+        }
         if (user.getRoles().contains(UserRole.ADMIN) && userRepository.countUserxByRolesContaining(UserRole.ADMIN) <= 1) {
-            //TODO Add log event here:
-            return false;
+            throw new LastAdminException();
         }
         userRepository.delete(user);
-        return true;
     }
 
 
@@ -95,33 +98,30 @@ public class UserxService {
     }
 
     @PreAuthorize("hasAuthority('ADMIN') or principal.username eq #username")
-    public Userx updateUser(String username, String firstname, String lastname, String email, Collection<UserRole> roles) {
-        if (roles.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+    public Userx updateUser(String username, String firstname, String lastname, String email, Collection<UserRole> roles) throws UserAlreadyExistsException {
+        if (userRepository.existsByEmail(email) && !userRepository.findByEmail(email).getUsername().equals(username)) {
+            throw new UserAlreadyExistsException();
         }
         Userx user = loadUser(username);
         user.setFirstName(firstname);
         user.setLastName(lastname);
         user.setEmail(email);
-        user.setRoles(Set.copyOf(roles));
+        user.setRoles(new HashSet<>(roles));
         return saveUser(user);
     }
 
     @PreAuthorize("hasAuthority('ADMIN')")
-    public Userx createUser(String username, String firstname, String lastname, String email, Collection<UserRole> roles, String password) {
-        if (roles.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+    public Userx createUser(String username, String firstname, String lastname, String email, Collection<UserRole> roles, String password) throws UserAlreadyExistsException {
+        if (userRepository.existsByUsername(username) || userRepository.existsByEmail(email)) {
+            throw new UserAlreadyExistsException();
         }
         Userx user = new Userx();
         user.setUsername(username);
         user.setFirstName(firstname);
         user.setLastName(lastname);
         user.setEmail(email);
-        user.setRoles(Set.copyOf(roles));
+        user.setRoles(new HashSet<>(roles));
         user.setPassword(WebSecurityConfig.passwordEncoder().encode(password));
         return saveUser(user);
     }
 }
-
-
-
